RAG-audio-support-assistant-main/requirements.txt
# Core dependencies for Audio RAG Support Assistant

# Speech-to-text
deepgram-sdk>=3.0.0

# LLM and embeddings
google-genai>=0.3.0
sentence-transformers>=2.2.0

# Vector database
chromadb>=1.5.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0

# Text processing
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: enhanced PDF processing
PyMuPDF>=1.23.0